Move shared utilities to the version independent shared package
Our versioned tests have a lot of versionless utilities, which can be moved in one shared place
and save us some maintenance burden, some investigation burden, etc.
